Reload AppUI routes on enter only.
